Removed old ItemList code and replaced it with the new stuff.
Renamed newitemlist to simply itemlist.  Deleted old itemlist and itemtrack.

Added some more placeholder sorts.  I believe we now have a class for every
sort we need, though many aren't implemented.
