+++ tv/lib/frontends/widgets/application.py
@@ -70,7 +70,7 @@ from miro.frontends.widgets import searchfilesdialog
 from miro.frontends.widgets import removefeeds
 from miro.frontends.widgets import diagnostics
 from miro.frontends.widgets import crashdialog
-from miro.frontends.widgets import newitemlist as itemlist
+from miro.frontends.widgets import itemlist
 from miro.frontends.widgets import itemlistcontroller
 from miro.frontends.widgets import prefpanel
 from miro.frontends.widgets import displays
